Chapter 23 Challenge 1 (Switch)
We have to discard any local variables created inside the loop in the
continue statement because 'continue' could happen to be on the last
iteration, causing the OP_POP commands created by endScope() to not be
executed, which causes problems later (e.g. when accessing the value of
local variable since the positions will be off).